# Add assembly for amd64 as well add Seed* funcs