--- a/murmur32_decl.go
+++ b/murmur32_decl.go
@@ -1,0 +1,5 @@
+// +build amd64
+
+package murmur3
+
+func Sum32(data []byte) (h1 uint32)