Each commit explains the reason for the commit. Hand assembly increases performance most on small inputs (minuscule little bit on large inputs) due to better branch layout, register usage, and conditional checks. Large inputs are mostly unaffected because the loop go compiles is pretty solid.