@twmb 2016-04-26
It appears that this gets only slightly faster than 1.4.2, the perf of which was gained from 1.3 and lost in 1.5 and 1.6. Go tip shows perf getting back to 1.4.2 levels, meaning the asm will provide minimal benefit going forward.

Because of this, and the overhead that extra code would entail, I'll close this PR in favor of a new PR adding only the middle two commits.