@spaolacci 2016-04-27
Ok. I was wondering as well about how large would the impact w.r.t tip. Out of curiosity, do you have those numbers to share?

I incidentally confirm there's interest for the `NewWithSeedxxx` constructors.